# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar archivos, directorios y etiquetas en Nostalgia.",
        "title": "Nostalgia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/directories/search": {
            "get": {
                "description": "Busca directorios por nombre y fecha",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "directories"
                ],
                "summary": "Buscar directorios",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Texto a buscar en el nombre",
                        "name": "contains",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha mínima (YYYY-MM-DD)",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha máxima (YYYY-MM-DD)",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Resultados por página",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NDirectorySearchResultDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/directories/{id}": {
            "get": {
                "description": "Devuelve los detalles de un directorio específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "directories"
                ],
                "summary": "Obtener directorio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Directory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NDirectoryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/directories/{id}/directories": {
            "get": {
                "description": "Devuelve la lista de subdirectorios dentro de un directorio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "directories"
                ],
                "summary": "Obtener subdirectorios de un directorio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Directory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.NDirectoryDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/directories/{id}/files": {
            "get": {
                "description": "Devuelve la lista de archivos dentro de un directorio",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "directories"
                ],
                "summary": "Obtener archivos de un directorio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Directory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.NFileDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/files/search": {
            "get": {
                "description": "Busca archivos por nombre, fecha y tipo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Buscar archivos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Texto a buscar en el nombre",
                        "name": "contains",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Fecha mínima (YYYY-MM-DD)",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha máxima (YYYY-MM-DD)",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "any",
                        "description": "Tipo de archivo (any, image, video)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Resultados por página",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NFileSearchResultDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "description": "Devuelve los detalles de un archivo específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Obtener archivo por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NFileDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.NDirectoryDto": {
            "type": "object",
            "required": [
                "full_path",
                "name"
            ],
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NDirectoryDto"
                    }
                },
                "directory_count": {
                    "type": "integer"
                },
                "file_count": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NFileDto"
                    }
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_directory_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "size_raw": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.NDirectorySearchResultDto": {
            "type": "object",
            "properties": {
                "contains": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NDirectoryDto"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.NFileDto": {
            "type": "object",
            "required": [
                "hash",
                "name",
                "path"
            ],
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_directory": {
                    "$ref": "#/definitions/dto.NDirectoryDto"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "size_raw": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.NFileSearchResultDto": {
            "type": "object",
            "properties": {
                "contains": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NFileDto"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}